refactor: PointConverter의 위치 변경 및 View에서 Point를 반환하도록 수정
