+++ src/main/java/chess/view/PointConverter.java
@@ -1,4 +1,4 @@
-package chess.controller;
+package chess.view;
 
 import java.awt.Point;
